[Oztechan/CCC#1880] Move shared classes into :client:core:shared
