+++ android/feature/mobile/src/main/kotlin/com/oztechan/ccc/android/feature/mobile/util/DeviceUtil.kt
@@ -9,7 +9,7 @@ import android.content.res.Configuration
 import android.content.res.Resources
 import android.os.Build
 import androidx.appcompat.app.AppCompatDelegate
-import com.oztechan.ccc.client.model.AppTheme
+import com.oztechan.ccc.client.core.shared.model.AppTheme
 import java.util.Locale
 
 fun updateBaseContextLocale(context: Context): Context? {
