+++ android/feature/mobile/src/main/kotlin/com/oztechan/ccc/android/feature/mobile/util/PurchaseUtil.kt
@@ -2,9 +2,9 @@ package com.oztechan.ccc.android.feature.mobile.util
 
 import com.oztechan.ccc.android.core.billing.model.ProductDetails
 import com.oztechan.ccc.android.core.billing.model.PurchaseHistoryRecord
-import com.oztechan.ccc.client.model.OldPurchase
-import com.oztechan.ccc.client.model.PremiumData
-import com.oztechan.ccc.client.model.PremiumType
+import com.oztechan.ccc.client.core.shared.model.OldPurchase
+import com.oztechan.ccc.client.core.shared.model.PremiumData
+import com.oztechan.ccc.client.core.shared.model.PremiumType
 
 internal fun List<ProductDetails>.toPremiumDataList(): List<PremiumData> = map {
     PremiumData(it.price, it.description, it.id)
