+++ client/core/shared/src/commonMain/kotlin/com/oztechan/ccc/client/core/shared/model/OldPurchase.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.client.model
+package com.oztechan.ccc.client.core.shared.model
 
 data class OldPurchase(
     val date: Long,
