+++ client/core/shared/src/commonMain/kotlin/com/oztechan/ccc/client/core/shared/model/PremiumData.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.client.model
+package com.oztechan.ccc.client.core.shared.model
 
 data class PremiumData(
     var cost: String,
