+++ client/core/shared/src/commonMain/kotlin/com/oztechan/ccc/client/core/shared/util/IndexUtil.kt
@@ -1,0 +1,5 @@
+package com.oztechan.ccc.client.core.shared.util
+
+fun Int.indexToNumber() = this + 1
+
+fun Int.numberToIndex() = this - 1
