+++ client/core/shared/src/commonTest/kotlin/com/oztechan/ccc/client/core/shared/model/AppThemeTest.kt
@@ -1,12 +1,11 @@
-package com.oztechan.ccc.client.model
+package com.oztechan.ccc.client.core.shared.model
 
 import com.oztechan.ccc.client.core.shared.Device
-import com.oztechan.ccc.client.helper.BaseTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
 
-internal class AppThemeTest : BaseTest() {
+internal class AppThemeTest {
     @Test
     fun getThemeByValue() {
         val appTheme = AppTheme.getThemeByValue(3)
