+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/settings/SettingsSEED.kt
@@ -1,10 +1,10 @@
 package com.oztechan.ccc.client.viewmodel.settings
 
+import com.oztechan.ccc.client.core.shared.model.AppTheme
 import com.oztechan.ccc.client.core.viewmodel.BaseData
 import com.oztechan.ccc.client.core.viewmodel.BaseEffect
 import com.oztechan.ccc.client.core.viewmodel.BaseEvent
 import com.oztechan.ccc.client.core.viewmodel.BaseState
-import com.oztechan.ccc.client.model.AppTheme
 
 // State
 data class SettingsState(
